style(*): rename modules to use lowercase
