+++ src/ramstk/controllers/hardware/matrixmanager.py
