--- src/ramstk/modules/hazops/__init__.py
@@ -1,3 +1,0 @@
-# RAMSTK Local Imports
-from .Controller import HazardAnalysisDataController as dtcHazardAnalysis
-from .Model import HazardAnalysisDataModel as dtmHazardAnalysis
